main class
Quite a messy main class. It declares all the planets and the moons of each planet and then goes into an infinite loop which makes all the planets orbit the sun. The first project I was incredibly proud of because I could see that I have improved. If I was to improve it now, I'd fix those declarations, they look horrible. I also feel that the infinite loop can be optimised more.